Send `remoteConnectionSuccess` when the remote connection succeeds
