+++ src/vs/workbench/services/remote/browser/remoteAgentService.ts
@@ -18,8 +18,6 @@ import { Registry } from 'vs/platform/registry/common/platform';
 import { IWorkbenchContribution, IWorkbenchContributionsRegistry, Extensions } from 'vs/workbench/common/contributions';
 import { IHostService } from 'vs/workbench/services/host/browser/host';
 import { LifecyclePhase } from 'vs/workbench/services/lifecycle/common/lifecycle';
-import { ITelemetryService } from 'vs/platform/telemetry/common/telemetry';
-import { getRemoteName } from 'vs/platform/remote/common/remoteHosts';
 
 export class RemoteAgentService extends AbstractRemoteAgentService implements IRemoteAgentService {
 
@@ -41,29 +39,10 @@ class RemoteConnectionFailureNotificationContribution implements IWorkbenchContr
 		@IRemoteAgentService remoteAgentService: IRemoteAgentService,
 		@IDialogService private readonly _dialogService: IDialogService,
 		@IHostService private readonly _hostService: IHostService,
-		@IWorkbenchEnvironmentService private readonly _environmentService: IWorkbenchEnvironmentService,
-		@ITelemetryService private readonly _telemetryService: ITelemetryService,
 	) {
 		// Let's cover the case where connecting to fetch the remote extension info fails
 		remoteAgentService.getRawEnvironment()
 			.then(undefined, (err) => {
-
-				type RemoteConnectionFailureClassification = {
-					web: { classification: 'SystemMetaData'; purpose: 'PerformanceAndHealth' };
-					remoteName: { classification: 'SystemMetaData'; purpose: 'PerformanceAndHealth' };
-					message: { classification: 'SystemMetaData'; purpose: 'PerformanceAndHealth' };
-				};
-				type RemoteConnectionFailureEvent = {
-					web: boolean;
-					remoteName: string | undefined;
-					message: string;
-				};
-				this._telemetryService.publicLog2<RemoteConnectionFailureEvent, RemoteConnectionFailureClassification>('remoteConnectionFailure', {
-					web: true,
-					remoteName: getRemoteName(this._environmentService.remoteAuthority),
-					message: err ? err.message : ''
-				});
-
 				if (!RemoteAuthorityResolverError.isHandled(err)) {
 					this._presentConnectionError(err);
 				}
